Replace os.path with Pathlib and fix matching_report
matching_report should be appending to the file (hence the
erase=True on the first call), but it actually just overwrites
the file each time it's called.

Also makes the default for the output_path parameter in match()
consistent with the default config value, and with the default
documented in the README "output" (which is the standard output
directory in an opensafely study repo.