Update workflow to publish latest tag manually
We don't intend to publish to pypi automatically anymore, but
there's a possibility we might want to do another release manually.
The build and publish workflow no longer automatically tags. Instead
it checks for the latest tag, and the latest pypi release, and will
publish a new release if they don't match.