Parse and validate config before calling match()
Adds a placeholder method for doing various validations on the
provided config, and doing any parsing/conversions that may be useful,
prior to loading the dataframes and starting any calculations.
Currently just ensures that an output_path string is converted to a
Path.

Currently there's a bunch of validation done during the matching
process, this will allow us to fail faster.