Allow control dataset to omit the index date column
Controls may not have an index date available. Users can choose to
provide a value for `generate_match_index_date`, which will populate
the control index date from their matched case. In this case, there's
no requirement for the control dataset to include an index date column
(and certainly no need to require users to generate control datasets
with placeholder index date columns just so that the matching process
can overwrite it)

(Note that a later commit will validate that an index date column is
present in the matching dataset if `generate_match_index_date` is
not provided)