Check and validate the config in match() if necessary
In case the match() function is used outside of the command line, we
make sure that it's been validated first.

When matching is a reusable action, we expect that it will always
be called with the __main__.py entrypoint, but it's possible that
someone could use it as a python library still.