fix: update the public module names
Update the publicly exposed module names and fix the README
accordingly.

I plan to do one last release before we move this repo and convert
it to a reusable action. Currently the documentation in the README
(on main) doesn't work with the latest released PyPI version, which
a user has already encountered. Once we deprecate the package, I'll
add a link in the README to the docs for the last PyPI release, but
at that point it'll be more obvious that the version on main isn't
intended for use as a locally installed package anymore.